+++ src/pytemplator/__init__.py
@@ -1,6 +1,6 @@
 """Top-level package for pytemplator."""
 
 
-__author__ = """Arnaud Blois"""
-__email__ = "a.blois@ucl.ac.uk"
-__version__ = "0.1.0a1"
+__author__ = "Arnaud Blois"
+__email__ = "hi@arnaudblois.com"
+__version__ = "0.1.0a2"